+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Changed
 
 - Updated gradle to 9.5.1
+- layout/index.html slight visual and performance improvements
 
 ### Fixed